refactor: don't expose CallbacksRegistry as an internal module
