+++ filenames.gypi
@@ -48,7 +48,6 @@
       'lib/browser/init.js',
       'lib/browser/objects-registry.js',
       'lib/browser/rpc-server.js',
-      'lib/common/api/callbacks-registry.js',
       'lib/common/api/clipboard.js',
       'lib/common/api/crash-reporter.js',
       'lib/common/api/deprecate.js',
@@ -63,6 +62,7 @@
       'lib/common/init.js',
       'lib/common/parse-features-string.js',
       'lib/common/reset-search-paths.js',
+      'lib/renderer/callbacks-registry.js',
       'lib/renderer/chrome-api.js',
       'lib/renderer/content-scripts-injector.js',
       'lib/renderer/init.js',
