+++ lib/common/api/module-list.js
@@ -5,7 +5,6 @@ module.exports = [
   {name: 'nativeImage', file: 'native-image'},
   {name: 'shell', file: 'shell'},
   // The internal modules, invisible unless you know their names.
-  {name: 'CallbacksRegistry', file: 'callbacks-registry', private: true},
   {name: 'deprecate', file: 'deprecate', private: true},
   {name: 'deprecations', file: 'deprecations', private: true},
   {name: 'isPromise', file: 'is-promise', private: true}
