+++ lib/renderer/api/remote.js
@@ -1,9 +1,10 @@
 'use strict'
 
 const v8Util = process.atomBinding('v8_util')
-const {ipcRenderer, isPromise, CallbacksRegistry} = require('electron')
+const {ipcRenderer, isPromise} = require('electron')
 const resolvePromise = Promise.resolve.bind(Promise)
 
+const CallbacksRegistry = require('../callbacks-registry')
 const bufferUtils = require('../../common/buffer-utils')
 
 const callbacksRegistry = new CallbacksRegistry()
