+++ lib/renderer/callbacks-registry.js
