+++ lib/sandboxed_renderer/api/module-list.js
@@ -2,11 +2,6 @@ const features = process.atomBinding('features')
 
 module.exports = [
   {
-    name: 'CallbacksRegistry',
-    load: () => require('../../common/api/callbacks-registry'),
-    private: true
-  },
-  {
     name: 'crashReporter',
     load: () => require('../../common/api/crash-reporter')
   },
